==> 00_unit_root.ipynb <==
## --------------------
## Author: Martin Řanda
## Year: 2023
##
## R Version 4.2.2
## Package versions:
## aTSA 3.1.2
## dplyr 1.0.10
## forecast 8.21
## librarian 1.8.1
## tseries 0.10-53
## xts 0.13.0
## zoo 1.8.11
##
## --------------------
if (!require(librarian)) install.packages("librarian")
librarian::shelf(dplyr, xts, aTSA, forecast, tseries)
# Set timezone to CET, locale to English
Sys.setenv(TZ = "CET")
Sys.setlocale("LC_TIME", "English")

series_raw <- readRDS("data/series_1h_2011_2021_xts.rds")
series_m_raw <- readRDS("minutes/data/load_1min_2011_2021_xts.rds")

### Unit root tests

Raw data, all variables.

aTSA::adf.test(series_m_raw, nlag = 10)


aTSA::adf.test(series_raw[, "load_mw"], nlag = 10)


aTSA::adf.test(series_raw[, "air_pressure"], nlag = 10)


aTSA::adf.test(series_raw[, "air_pressure_diff"], nlag = 10)


aTSA::adf.test(series_raw[, "price_eur_mwh"], nlag = 10)


aTSA::adf.test(series_raw[, "temperature"], nlag = 10)


aTSA::adf.test(series_raw[, "temperature_diff"], nlag = 10)


aTSA::adf.test(series_raw[, "visibility_distance"], nlag = 10)


aTSA::adf.test(series_raw[, "visibility_distance_diff"], nlag = 10)


aTSA::adf.test(series_raw[, "wind_speed"], nlag = 10)


aTSA::adf.test(series_raw[, "wind_speed_diff"], nlag = 10)

***

### Run on differenced data

Only those for which the null hypothesis was not rejected in the majority of the cases

aTSA::adf.test(diff(series_m_raw), nlag = 10)

aTSA::adf.test(diff(series_raw[, "air_pressure"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "air_pressure_diff"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "price_eur_mwh"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "temperature"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "visibility_distance"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "visibility_distance_diff"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "wind_speed"]), nlag = 10)


aTSA::adf.test(diff(series_raw[, "wind_speed_diff"]), nlag = 10)

### Differences with `tseries::adf.test` default values

tseries::adf.test(na.omit(diff(series_m_raw)))

tseries::adf.test(na.omit(diff(series_raw[, "air_pressure"])))

tseries::adf.test(na.omit(diff(series_raw[, "air_pressure_diff"])))

tseries::adf.test(na.omit(diff(series_raw[, "price_eur_mwh"])))

tseries::adf.test(na.omit(diff(series_raw[, "temperature"])))

tseries::adf.test(na.omit(diff(series_raw[, "visibility_distance"])))

tseries::adf.test(na.omit(diff(series_raw[, "visibility_distance_diff"])))

tseries::adf.test(na.omit(diff(series_raw[, "wind_speed"])))

tseries::adf.test(na.omit(diff(series_raw[, "wind_speed_diff"])))
